+++ R/tq_portfolio.R
@@ -63,7 +63,7 @@
 #' # Get returns for individual stock components
 #' monthly_returns_stocks <- FANG %>%
 #'     group_by(symbol) %>%
-#'     tq_transmute(Ad, periodReturn, period = "monthly")
+#'     tq_transmute(adjusted, periodReturn, period = "monthly")
 #'
 #' ##### Portfolio Aggregation Methods #####
 #'