+++ tests/testthat/test_tq_performance.R
@@ -10,14 +10,14 @@ Ra <- c("AAPL", "GOOG", "NFLX") %>%
            from = "2010-01-01",
            to   = "2015-12-31") %>%
     group_by(symbol) %>%
-    tq_transmute(Ad, periodReturn, period = "monthly", col_rename = "Ra")
+    tq_transmute(adjusted, periodReturn, period = "monthly", col_rename = "Ra")
 
 # Get returns for SP500 as baseline
 Rb <- "^GSPC" %>%
     tq_get(get  = "stock.prices",
            from = "2010-01-01",
            to   = "2015-12-31") %>%
-    tq_transmute(Ad, periodReturn, period = "monthly", col_rename = "Rb")
+    tq_transmute(adjusted, periodReturn, period = "monthly", col_rename = "Rb")
 
 # Merge stock returns with baseline
 RaRb <- left_join(Ra, Rb, by = c("date" = "date"))